requirements.txt
Django==1.6.5
django-modeltranslation==0.7.3
south==1.0